Extract interface for IIIFAuthBuilder to make introducing v2 easier
